LayoutTests/imported/w3c/web-platform-tests/css/css-values/attr-length-specified-expected.txt


PASS type(<length>) lowercase
FAIL type(<length>) uppercase assert_equals: expected "3em" but got "3EM"
PASS Unit (em) number
